+++ SRC/Classes/psp.class.pspuser.ps1
@@ -6,16 +6,22 @@ class PSPUser {
     [datetime]$DOB
     [String]$SamAccountName
     [String]$PlainTextPassword
+    [String]$ID
 
     #Constructor
-    PSPUser ([String]$FirstName, [String]$LastName, [String]$DOB) {
+    PSPUser ([String]$FirstName, [String]$LastName, [String]$DOB, [String]$ID) {
         $this.FirstName = $FirstName
         $this.LastName = $LastName
         $this.DOB = $DOB
+        $this.ID = $ID
     }
 
     #Set Username
     SetUsername ([string]$username) {
         $this.SamAccountName = $username
+    }
+    
+    SetPlainTextPassword ([string]$PlainTextPassword) {
+        $this.PlainTextPassword = $PlainTextPassword
     }
 }